[chore #158052672] API: restructure, add blueprints
