+++ api/auth/__init__.py
@@ -1,0 +1,6 @@
+#api/auth/__init__.py
+from flask import Blueprint
+
+auth = Blueprint('auth', __name__)
+
+from . import views
