+++ api/mrequests/__init__.py
@@ -1,0 +1,7 @@
+#api/requests/__init__.py
+
+from flask import Blueprint
+
+mrequests = Blueprint('mrequests', __name__)
+
+from . import views
