+++ api/tests/test_auth.py
@@ -8,7 +8,7 @@ This module include tests to the auth endpoints of the API ie
 import unittest
 import json
 from api import create_app
-from api.app.models import User, Request
+from api.models import User, Request
 
 
 class TestAPIAuth(unittest.TestCase):
